requirements.txt
ajsonrpc==1.2.0
alembic==1.16.1
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
apispec==6.8.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.4
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
cors==1.0.1
deep-translator==1.11.4
distlib==0.3.9
Django==5.2.2
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.16.0
dotenv==0.9.9
filelock==3.18.0
flasgger==0.9.7.1
Flask==2.1.3
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-smorest==0.42.3
Flask-SQLAlchemy==2.5.1
flatbuffers==25.2.10
future==1.0.0
gevent==25.5.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
googletrans-py==4.0.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idf-component-manager==2.2.2
idna==2.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mistune==3.1.3
mpmath==1.3.0
murf==1.2.2
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
packaging==25.0
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
platformio==6.1.18
pooch==1.8.2
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyelftools==0.32
PyJWT==2.10.1
PyMatting==1.1.14
pyparsing==3.2.3
pypiwin32==223
pyreadline3==3.5.4
pyserial==3.5
PySocks==1.7.1
python-dotenv==1.1.0
pyttsx3==2.90
pytz==2025.2
pywin32==310
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rembg==2.0.67
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.24.0
rq==2.3.3
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scipy==1.16.0
semantic-version==2.10.0
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.46
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.6.11
tldextract==5.3.0
tqdm==4.67.1
truststore==0.10.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
virtualenv==20.30.0
webargs==8.7.0
Werkzeug==2.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
